Create MyInfoSmallQR.py
the main file. run this in any IDE with your info inserted in the spaces and with the filepath of your choosing. NOTE: it is optional to remove the URIs from the datapoints, so you can remove 'tel:' and 'mailto:', though i prefer to keep them because why not, can't hurt.